Cash for ApplicationContainer and ProvideService
