+++ RedCatEngineUnityProject/Packages/DependencyInjection/Containers/Interfaces/IApplicationContainer.cs
@@ -1,0 +1,6 @@
+namespace RedCatEngine.DependencyInjection.Containers.Interfaces
+{
+	public interface IApplicationContainer : IBinderApplicationContainer, IGetterApplicationContainer
+	{
+	}
+}
