+++ RedCatEngineUnityProject/Packages/DependencyInjection/Containers/Interfaces/IGetterApplicationContainer.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f891f8de43834d349b52dd6e2136b4c7
+timeCreated: 1719517743
