+++ RedCatEngineUnityProject/Packages/DependencyInjection/Specials.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 82dba7ee8f074b96a841d90bd4a940b7
+timeCreated: 1719516693
