+++ RedCatEngineUnityProject/Packages/DependencyInjection/Tests/ServiceLocatorGettersTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using NUnit.Framework;
 using RedCatEngine.DependencyInjection.Containers;
+using RedCatEngine.DependencyInjection.Containers.Interfaces;
 using RedCatEngine.DependencyInjection.Tests.SpecialSubClasses;
 
 namespace RedCatEngine.DependencyInjection.Tests
@@ -12,7 +13,7 @@ namespace RedCatEngine.DependencyInjection.Tests
 		[SetUp]
 		public void SetUp()
 		{
-			_applicationContainer = new ServiceLocatorApplicationContainer();
+			_applicationContainer = new ApplicationContainer();
 		}
 
 		[Test]
